fix(ci): 修复 tsc/dprint/clippy 错误 — i18n、类型、格式化
- GlobalErrorBoundary: 添加 retryKey 到 state 接口并实现增量逻辑
- codeExecutor: 使用 PYODIDE_SRI integrity hash
- 所有 i18n 修正: dprint 格式化 + locale 文件
- 移除误提交的临时文件

+++ src-tauri/crates/core/src/repo/gateway_request_log.rs
@@ -14,8 +14,8 @@ pub async fn record_request_log(
     provider_id: Option<&str>,
     status_code: i32,
     duration_ms: i32,
-    request_tokens: i32,
-    response_tokens: i32,
+    request_tokens: i64,
+    response_tokens: i64,
     error_message: Option<&str>,
 ) -> Result<()> {
     gateway_request_logs::ActiveModel {
